+++ src/OpenTelemetry.Resources.Azure/CHANGELOG.md
@@ -17,6 +17,9 @@ Released 2026-Sep-10
   * `azure_vm` to `azure.vm`
   ([#5143](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5143))
 
+* Detect Azure Functions environments before Azure App Service and emit Azure
+  Functions resource attributes.
+
 * Log an event when a resource detector does not detect any attributes because
   none of the environment variables that identify the platform are set.
   ([#5169](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5169))